Move `MemoizingObserver` to the package level
