+++ server/src/test/java/io/spine/server/QueryServiceShould.java
@@ -25,8 +25,8 @@ import io.spine.client.QueryResponse;
 import io.spine.core.EventContext;
 import io.spine.core.Responses;
 import io.spine.core.Subscribe;
+import io.spine.grpc.MemoizingObserver;
 import io.spine.grpc.StreamObservers;
-import io.spine.grpc.StreamObservers.MemoizingObserver;
 import io.spine.server.projection.Projection;
 import io.spine.server.projection.ProjectionRepository;
 import io.spine.server.stand.Stand;
