+++ server/src/test/java/io/spine/server/commandbus/AbstractCommandBusTestSuite.java
@@ -34,8 +34,8 @@ import io.spine.core.CommandValidationError;
 import io.spine.core.IsSent;
 import io.spine.core.Status;
 import io.spine.core.TenantId;
+import io.spine.grpc.MemoizingObserver;
 import io.spine.grpc.StreamObservers;
-import io.spine.grpc.StreamObservers.MemoizingObserver;
 import io.spine.server.command.Assign;
 import io.spine.server.command.CommandHandler;
 import io.spine.server.commandstore.CommandStore;
