+++ server/src/test/java/io/spine/server/commandbus/SingleTenantCommandBusShould.java
@@ -28,7 +28,7 @@ import io.spine.core.CommandEnvelope;
 import io.spine.core.CommandValidationError;
 import io.spine.core.Failure;
 import io.spine.core.IsSent;
-import io.spine.grpc.StreamObservers.MemoizingObserver;
+import io.spine.grpc.MemoizingObserver;
 import io.spine.server.bus.EnvelopeValidator;
 import io.spine.server.command.Assign;
 import io.spine.server.command.CommandHandler;
